Move test to more appropriate class
